+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 - The config command, to view and change the config through the CLI.
 - Support for using external files for test scripts in the package metadata.
+- The fuzzy-search feature, to provide a package suggestion in case of a wrong package parameter.
 
 ### Removed
 - The repositories command, this command is now integrated in the new config command.